requirements.txt
asgiref==3.3.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.4
cryptography==3.3.1
Django==3.1.11
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-split-settings==0.2.5
djangorestframework==3.12.4
enum34==1.1.6
Markdown==3.3.4
Pillow==8.2.0
pycparser==2.17
pytz==2020.4
six==1.10.0
soupsieve==2.2.1
sqlparse==0.4.1
waitress==1.4.4
whitenoise==5.2.0
wincertstore==0.2
